move scheduler_utils::matmul_utils into mma_util
