+++ csrc/scheduler/utils.cpp
@@ -1564,337 +1564,6 @@ BroadcastMultipleInformation getBroadcastMultiples(
   return bcast_info;
 }
 
-namespace matmul_utils {
-
-void scheduleWarpTileWithReduction(TensorView* tv, MatMulTileOptions tile) {
-  // Assumes
-  // [M, N, K]
-  auto cta_tile = tile.cta_tile;
-  auto warp_tile = tile.warp_tile;
-  auto instruction_tile = tile.instruction_tile;
-
-  TORCH_CHECK(
-      cta_tile.k % warp_tile.k == 0,
-      "Number of warp on k dimension need to be integer");
-
-  int num_warp_k = cta_tile.k / warp_tile.k;
-
-  mma_util::checkDimSize(
-      tv, {-3, -2, -1}, {cta_tile.m, cta_tile.n, cta_tile.k});
-
-  if (num_warp_k == 1) {
-    // Non split K over warp case:
-
-    //       -3   -2  -1
-    //[...    M,   N,  K]
-    // Distribute warp tile:
-    tv->split(-3, warp_tile.m);
-    tv->split(-2, warp_tile.n);
-
-    //  -5   -4   -3   -2   -1
-    // [Mwo  Mw  Nwo   Nw   K]
-    tv->split(-4, instruction_tile.m);
-    tv->split(-2, instruction_tile.n);
-    tv->split(-1, instruction_tile.k);
-
-    //   -8  -7 -6 -5 -4 -3  -2 -1
-    // [Mwo Mw Mi Nwo Nw Ni Kwo Ki]
-
-    tv->reorder({{-7, -5}, {-6, -3}, {-5, -6}, {-3, -2}, {-2, -8}, {-8, -7}});
-    //   -8  -7 -6  -5 -4 -3 -2 -1
-    // [Kwo Mwo Nwo Mw Nw Mi Ni Ki]
-  } else {
-    // Split K over warp case:
-    // Main difference is that an additional
-    //  thread dimension needs to be reserved
-    //  for cross warp reduction:
-    //       -3   -2  -1
-    //[...    M,   N,  K]
-    // Distribute warp tile:
-    tv->split(-3, warp_tile.m);
-    tv->split(-2, warp_tile.n);
-    tv->split(-1, warp_tile.k);
-
-    //   -6  -5   -4   -3   -2   -1
-    // [Mwo  Mw  Nwo   Nw   Kwo  Kw]
-    tv->split(-5, instruction_tile.m);
-    tv->split(-3, instruction_tile.n);
-    tv->split(-1, instruction_tile.k);
-
-    //  -9  -8  -7 -6 -5 -4 -3 -2 -1
-    // [Mwo Mw Mi Nwo Nw Ni Kwo Kw Ki]
-
-    tv->reorder({{-8, -6}, {-7, -3}, {-6, -8}, {-4, -2}, {-3, -7}, {-2, -4}});
-    //  -9   -8  -7 -6 -5 -4 -3 -2 -1
-    // [Mwo  Nwo Ko Mw Nw Kw, Mi Ni Ki]
-
-    tv->merge(-9);
-    //  -8  -7 -6 -5 -4   -3 -2 -1
-    // [MNwo Ko Mw Nw Kw, Mi Ni Ki]
-  }
-}
-
-void scheduleWarpTileWithNoReduction(TensorView* tv, MatMulTileOptions tile) {
-  // Assumes
-  // [M, N, K]
-  auto cta_tile = tile.cta_tile;
-  auto warp_tile = tile.warp_tile;
-  auto instruction_tile = tile.instruction_tile;
-
-  mma_util::checkDimSize(tv, {-2, -1}, {cta_tile.m, cta_tile.n});
-
-  TORCH_CHECK(
-      cta_tile.k % warp_tile.k == 0,
-      "Number of warp on k dimension need to be integer");
-
-  int num_warp_k = cta_tile.k / warp_tile.k;
-
-  //        -2  -1
-  //[...    M,   N]
-
-  // Distribute warp tile:
-  tv->split(-2, warp_tile.m);
-  tv->split(-1, warp_tile.n);
-
-  //  -4   -3   -2   -1
-  // [Mwo  Mw  Nwo   Nw ]
-  tv->split(-3, instruction_tile.m);
-  tv->split(-1, instruction_tile.n);
-
-  //  -6 -5  -4 -3 -2 -1
-  // [Mwo Mw Mi Nwo Nw Ni]
-
-  tv->reorder({{-5, -4}, {-4, -2}, {-3, -5}, {-2, -3}});
-
-  //  -6   -5  -4 -3 -2 -1
-  // [Mwo  Nwo Mw Nw Mi Ni]
-
-  if (num_warp_k != 1) {
-    // The non reduction warps are merged together
-    //  to save one thread dim for cross dim reduce.
-    tv->merge(-6);
-    //  -5  -4 -3 -2 -1
-    // [MNo Mw Nw Mi Ni]
-  }
-}
-
-//! Split the innermost dim to a vectorized load
-void scheduleContiguousVectorLoad(
-    TensorView* tv,
-    MatMulTileOptions tile,
-    int vector_word,
-    bool vectorize) {
-  auto warp_dims = tile.cta_tile / tile.warp_tile;
-  int num_of_thread = warp_dims.m * warp_dims.n * warp_dims.k * 32;
-
-  tv->split(-1, num_of_thread * vector_word);
-  tv->split(-1, vector_word);
-  // [..., thread, vec]
-  // distribute to warp: for tidx
-  tv->split(-2, 32);
-
-  //      -3    -2    -1
-  // [...warp, lane, vec]
-
-  if (warp_dims.k == 1) {
-    //      -4     -3    -2    -1
-    // [...warpM, warpN, lane, vec]
-    tv->split(-3, warp_dims.n);
-  } else {
-    //      -4     -3    -2    -1
-    // [...warpMN, warpR, lane, vec]
-    tv->split(-3, warp_dims.k);
-  }
-
-  if (vectorize) {
-    tv->axis(-1)->parallelize(ParallelType::Vectorize);
-  }
-
-  tv->axis(-2)->parallelize(ParallelType::TIDx);
-  tv->axis(-3)->parallelize(ParallelType::TIDy);
-  tv->axis(-4)->parallelize(ParallelType::TIDz);
-}
-
-void makeTile(TensorView* tv, std::vector<int> tile_sizes) {
-  TORCH_CHECK(
-      tv->domain()->domain().size() >= tile_sizes.size(),
-      "Tensor dimension less than tile dimension!");
-
-  // Number of inner dimensions we are tiling.
-  const auto tile_dimension_size = tile_sizes.size();
-
-  // Split the inner dimensions:
-  for (auto idx : c10::irange(tile_dimension_size)) {
-    // Using negative indexing to accomodate potential batching
-    //  dimensions on the further left. Eg.:
-    //  0, 1, 2   ->         -3,-2,-1
-    // [M, N, K]  -> [B0, B1, M, N, K]
-    tv->split(idx - tile_dimension_size, tile_sizes.at(idx));
-  }
-
-  // The transformation happened should look like:
-  //   Before               After
-  // [..., M, N, K] -> [..., Mo, Mi, No, Ni, Ko, Ki]
-
-  // Re-order the tiles so that all the outer tiles are
-  //  on the left of all the inner tiles
-  std::unordered_map<int, int> reorder_map_old_to_new;
-
-  // Number of tiled inner dimensions after we split.
-  const auto split_tile_dimension_size = 2 * tile_dimension_size;
-  for (auto idx : c10::irange(split_tile_dimension_size)) {
-    // We want to reorder as follows:
-    //           Before
-    //
-    // [..., Mo, Mi, No, Ni, Ko, Ki] ->
-    //                 After
-    //      vvv group0 vvv  vvv group1 vvv
-    // [..., Mo, No, Ko,     Mi, Ni, Ki]
-
-    // The index offset within group of current
-    //  iterdomain, with grouping specified above.
-    auto index_within_group = idx / 2;
-
-    // The index of the group the current id belongs
-    //  to, as specified above.
-    auto group_index = idx % 2;
-
-    // Calculate the actual index after reordering
-    auto index_after_reorder =
-        group_index * tile_dimension_size + index_within_group;
-
-    // Add pair {idx_before, idx_after} to re-order map.
-    reorder_map_old_to_new.insert(std::make_pair(
-        idx - split_tile_dimension_size,
-        index_after_reorder - split_tile_dimension_size));
-  }
-
-  // Apply the re-order map to tensor
-  tv->reorder(reorder_map_old_to_new);
-}
-
-namespace {
-
-c10::optional<IterDomain*> getMaybeRootIfInnermostTiled(
-    IterDomain* id,
-    const std::unordered_set<IterDomain*>& maybe_rfactor_id_set) {
-  // Root id defaults to an "innermost id".
-  while (id->definition() && !maybe_rfactor_id_set.count(id)) {
-    if (auto split = dynamic_cast<Split*>(id->definition())) {
-      if (id == split->inner()) {
-        id = split->in();
-        continue;
-      }
-    }
-    // Didn't pass the inner most check, return empty.
-    return c10::nullopt;
-  }
-
-  return id;
-}
-
-} // namespace
-
-void orderTiledConcreteIdAsRoot(TensorView* tv) {
-  auto ndims = tv->nDims();
-
-  // Keep track of the left most position where we will
-  //  be reordering the axes.
-  auto leftmost_pos = ndims;
-
-  // Pull the root id's of the given tv.
-  std::unordered_set<IterDomain*> maybe_rfactor_id_set{
-      tv->getMaybeRFactorDomain().begin(), tv->getMaybeRFactorDomain().end()};
-
-  // Keep track of leaf positions that is either a reduction
-  //  or a broadcast.
-  // Note: Currently don't really see a case where this function
-  //  should be called on a reduction output tv, but adding them
-  //  here for completeness.
-  std::deque<int> broadcast_or_reduction_pos;
-
-  // Map the root id's to their innermost concrete id's
-  //  on the leaf.
-  std::unordered_map<IterDomain*, int> root_id_to_inner_leaf_pos;
-
-  // Try to re-order inner iterdomains from the innermost
-  //  position backward. This utility only tries to re-order
-  //  inner tiles on the innermost positions, like the resulting
-  //  tensor from makeTile utility.
-  // The re-ordering would first try to decide the inner iterdomains
-  //  we want to re-order. For this we start from the innermost position
-  //  and move back and collect all the iterdomains that we know
-  //  are inner tiles of some root domain or broadcast/reduction domains
-  //  that won't affect the concrete id layout.
-  // The collection process would stop whenever a iterdomain that is
-  //  neither an inner tile nor reduction/broadcast is found, and would
-  //  not re-order any iterdomain beyond that point to keep the
-  //  outer loop structure unchanged.
-  for (int64_t i = static_cast<int64_t>(ndims) - 1; i >= 0; i--) {
-    auto leaf_id = tv->axis(i);
-    if (leaf_id->isBroadcast() || leaf_id->isReduction()) {
-      // Register this reduction or broadcast axis
-      //  to reorder.
-      broadcast_or_reduction_pos.push_front(i);
-      leftmost_pos = i;
-      continue;
-    }
-    auto maybe_root =
-        getMaybeRootIfInnermostTiled(leaf_id, maybe_rfactor_id_set);
-
-    if (maybe_root.has_value()) {
-      // Found an innermost id, add them to the
-      //  axes to reorder.
-      TORCH_INTERNAL_ASSERT(
-          root_id_to_inner_leaf_pos
-              .insert(std::make_pair(maybe_root.value(), i))
-              .second,
-          "Multiple \"innermost\" id seen for root id :",
-          maybe_root.value()->toString(),
-          " on ",
-          tv->toString(),
-          " very likely an invariant is broken.");
-      leftmost_pos = i;
-    } else {
-      break;
-    }
-  }
-
-  // Calculate the ordering:
-
-  // pointer to the current target postion after
-  //  repordering
-  int current_pos = leftmost_pos;
-  std::unordered_map<int, int> reorder_map_old_to_new;
-
-  // first place all the broadcast and reduction on the left:
-  for (auto original_broadcast_or_reduction_pos : broadcast_or_reduction_pos) {
-    reorder_map_old_to_new[original_broadcast_or_reduction_pos] = current_pos++;
-  }
-
-  // Next put all the innermost leaf id's, we make sure that
-  //  the inner tile ordering follows the corresponding root
-  //  domain ordering by iterating on the root domain and
-  //  find their corresponding inner tile iterdomains from
-  //  the populated root_id_to_inner_leaf_pos.
-  for (auto root_id : tv->getMaybeRFactorDomain()) {
-    auto leaf_id_pos_it = root_id_to_inner_leaf_pos.find(root_id);
-    if (leaf_id_pos_it != root_id_to_inner_leaf_pos.end()) {
-      reorder_map_old_to_new[leaf_id_pos_it->second] = current_pos++;
-    }
-  }
-
-  // Validate that we have processed all inner ids or broadcast/reduction
-  //  ids we have registered.
-  TORCH_INTERNAL_ASSERT(
-      current_pos == (int)ndims, "Inconsistent ordering logic");
-
-  // Apply the new order:
-  tv->reorder(reorder_map_old_to_new);
-}
-
-} // namespace matmul_utils
-
 //! Propagate current transformations on from_tv to all graphs
 void transformPropagateToAllFrom(TensorView* from_tv, int pos) {
   TransformPropagator propagator(from_tv, pos);
