+++ test/test_gpu_tensorcore.cpp
@@ -29,6 +29,7 @@
 #include <root_domain_map.h>
 #include <scheduler/all_schedulers.h>
 #include <scheduler/matmul.h>
+#include <scheduler/mma_utils.h>
 #include <scheduler/reduction_utils.h>
 #include <scheduler/utils.h>
 #include <test/test_gpu_validator.h>
@@ -1019,10 +1020,8 @@ TEST_F(NVFuserTest, FusionMatmulMatmulAmpere_CUDA) {
   tv2r->computeAt(tv4c, 3);
 
   // Make warp tile
-  scheduler_utils::matmul_utils::scheduleWarpTileWithReduction(
-      tv4c, gemm_tile2);
-  scheduler_utils::matmul_utils::scheduleWarpTileWithNoReduction(
-      tv4, gemm_tile2);
+  mma_util::scheduleWarpTileWithReduction(tv4c, gemm_tile2);
+  mma_util::scheduleWarpTileWithNoReduction(tv4, gemm_tile2);
   //           -8   -7  -6 -5 -4 -3 -2 -1
   // [Mo No Ko Kwo Mwo Nwo Mw Nw Mi Ni Ki]
   tv3cr->computeAt(tv4c, -4);
@@ -1035,10 +1034,8 @@ TEST_F(NVFuserTest, FusionMatmulMatmulAmpere_CUDA) {
   tv2r->merge(-2);
 
   // [No,Ko,i,wy,wx,v]
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv2cw, gemm_tile2, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv2r, gemm_tile2, 8);
+  mma_util::scheduleContiguousVectorLoad(tv2cw, gemm_tile2, 8);
+  mma_util::scheduleContiguousVectorLoad(tv2r, gemm_tile2, 8);
   tv2cw->setMemoryType(MemoryType::Shared);
 
   // Schedule tv2 gmem read and smem write:
@@ -1083,10 +1080,8 @@ TEST_F(NVFuserTest, FusionMatmulMatmulAmpere_CUDA) {
 
   // Make warp tile:
   // -------------------------------------------------------------------------
-  scheduler_utils::matmul_utils::scheduleWarpTileWithReduction(
-      tv3c, gemm_tile1);
-  scheduler_utils::matmul_utils::scheduleWarpTileWithNoReduction(
-      tv3cw, gemm_tile1);
+  mma_util::scheduleWarpTileWithReduction(tv3c, gemm_tile1);
+  mma_util::scheduleWarpTileWithNoReduction(tv3cw, gemm_tile1);
 
   tv0cr->computeAt(tv3c, -4);
   tv1cr->computeAt(tv3c, -4);
@@ -1098,10 +1093,8 @@ TEST_F(NVFuserTest, FusionMatmulMatmulAmpere_CUDA) {
   // [Mo,Ko,M,K]
   tv0cw->merge(-2);
   tv0r->merge(-2);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0cw, gemm_tile1, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0r, gemm_tile1, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0cw, gemm_tile1, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0r, gemm_tile1, 8);
   tv0cw->setMemoryType(MemoryType::Shared);
   // [Mo,Ko,i,wy,wx,v]
 
@@ -1109,10 +1102,8 @@ TEST_F(NVFuserTest, FusionMatmulMatmulAmpere_CUDA) {
   tv1cw->merge(-2);
   tv1r->merge(-2);
   // [No,Ko,i,wy,wx,v]
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1cw, gemm_tile1, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1r, gemm_tile1, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1cw, gemm_tile1, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1r, gemm_tile1, 8);
   tv1cw->setMemoryType(MemoryType::Shared);
 
   // Schedule mma input
@@ -1326,9 +1317,8 @@ TEST_F(NVFuserTest, FusionMatmulSoftmaxMatmulAmpere_CUDA) {
   tv2r->computeAt(tv4c, 3);
 
   // Make warp tile
-  scheduler_utils::matmul_utils::scheduleWarpTileWithReduction(tv4c, gemm_tile);
-  scheduler_utils::matmul_utils::scheduleWarpTileWithNoReduction(
-      tv4, gemm_tile);
+  mma_util::scheduleWarpTileWithReduction(tv4c, gemm_tile);
+  mma_util::scheduleWarpTileWithNoReduction(tv4, gemm_tile);
   //           -8  -7  -6  -5 -4 -3 -2 -1
   // [Mo No Ko Kwo Mwo Nwo Mw Nw Mi Ni Ki]
   tv3cr->computeAt(tv4c, -4);
@@ -1341,10 +1331,8 @@ TEST_F(NVFuserTest, FusionMatmulSoftmaxMatmulAmpere_CUDA) {
   tv2r->merge(-2);
 
   // [No,Ko,i,wy,wx,v]
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv2cw, gemm_tile, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv2r, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv2cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv2r, gemm_tile, 8);
   tv2cw->setMemoryType(MemoryType::Shared);
 
   // Schedule tv2 gmem read and smem write:
@@ -1396,9 +1384,8 @@ TEST_F(NVFuserTest, FusionMatmulSoftmaxMatmulAmpere_CUDA) {
 
   // Make warp tile:
   // -------------------------------------------------------------------------
-  scheduler_utils::matmul_utils::scheduleWarpTileWithReduction(tv3c, gemm_tile);
-  scheduler_utils::matmul_utils::scheduleWarpTileWithNoReduction(
-      tv3, gemm_tile);
+  mma_util::scheduleWarpTileWithReduction(tv3c, gemm_tile);
+  mma_util::scheduleWarpTileWithNoReduction(tv3, gemm_tile);
 
   tv0cr->computeAt(tv3c, -4);
   tv1cr->computeAt(tv3c, -4);
@@ -1410,10 +1397,8 @@ TEST_F(NVFuserTest, FusionMatmulSoftmaxMatmulAmpere_CUDA) {
   // [Mo,Ko,M,K]
   tv0cw->merge(-2);
   tv0r->merge(-2);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0cw, gemm_tile, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0r, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0r, gemm_tile, 8);
   tv0cw->setMemoryType(MemoryType::Shared);
   // [Mo,Ko,i,wy,wx,v]
 
@@ -1421,10 +1406,8 @@ TEST_F(NVFuserTest, FusionMatmulSoftmaxMatmulAmpere_CUDA) {
   tv1cw->merge(-2);
   tv1r->merge(-2);
   // [No,Ko,i,wy,wx,v]
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1cw, gemm_tile, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1r, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1r, gemm_tile, 8);
   tv1cw->setMemoryType(MemoryType::Shared);
 
   // Schedule mma input
@@ -1902,9 +1885,8 @@ TEST_F(NVFuserTest, FusionAmpereMatmulTNcpAsync_CUDA) {
 
   // Make warp tile:
   // -------------------------------------------------------------------------
-  scheduler_utils::matmul_utils::scheduleWarpTileWithReduction(tv2c, gemm_tile);
-  scheduler_utils::matmul_utils::scheduleWarpTileWithNoReduction(
-      tv2, gemm_tile);
+  mma_util::scheduleWarpTileWithReduction(tv2c, gemm_tile);
+  mma_util::scheduleWarpTileWithNoReduction(tv2, gemm_tile);
   //           -8  -7  -6  -5 -4 -3 -2 -1
   // [Mo No Ko Kwo Mwo Nwo Mw Nw Mi Ni Ki]
   tv0cr->computeAt(tv2c, -4);
@@ -1914,16 +1896,14 @@ TEST_F(NVFuserTest, FusionAmpereMatmulTNcpAsync_CUDA) {
   // ---------------------------------------------------------------------------
   // [Mo,Ko,M,K]
   tv0cw->merge(-2);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0cw, gemm_tile, 8);
   tv0cw->setMemoryType(MemoryType::Shared);
   // [Mo,Ko,i,wy,wx,v]
 
   // [No,Ko,N,K]
   tv1cw->merge(-2);
   // [No,Ko,i,wy,wx,v]
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1cw, gemm_tile, 8);
   tv1cw->setMemoryType(MemoryType::Shared);
   // Schedule mma input
   // ---------------------------------------------------------------------------
@@ -2062,9 +2042,8 @@ TEST_F(NVFuserTest, FusionAmpereStridedBatchedMatmulTN_CUDA) {
 
   // Make warp tile:
   // -------------------------------------------------------------------------
-  scheduler_utils::matmul_utils::scheduleWarpTileWithReduction(tv2c, gemm_tile);
-  scheduler_utils::matmul_utils::scheduleWarpTileWithNoReduction(
-      tv2, gemm_tile);
+  mma_util::scheduleWarpTileWithReduction(tv2c, gemm_tile);
+  mma_util::scheduleWarpTileWithNoReduction(tv2, gemm_tile);
   //           -8  -7  -6  -5 -4 -3 -2 -1
   // [Mo No Ko Kwo Mwo Nwo Mw Nw Mi Ni Ki]
   tv0cr->computeAt(tv2c, -4);
@@ -2075,10 +2054,8 @@ TEST_F(NVFuserTest, FusionAmpereStridedBatchedMatmulTN_CUDA) {
   // [Mo,Ko,M,K]
   tv0cw->merge(-2);
   tv0r->merge(-2);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0cw, gemm_tile, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0r, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0r, gemm_tile, 8);
   tv0cw->setMemoryType(MemoryType::Shared);
   // [Mo,Ko,i,wy,wx,v]
 
@@ -2086,10 +2063,8 @@ TEST_F(NVFuserTest, FusionAmpereStridedBatchedMatmulTN_CUDA) {
   tv1cw->merge(-2);
   tv1r->merge(-2);
   // [No,Ko,i,wy,wx,v]
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1cw, gemm_tile, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1r, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1r, gemm_tile, 8);
   tv1cw->setMemoryType(MemoryType::Shared);
   // Schedule mma input
   // ---------------------------------------------------------------------------
@@ -2229,9 +2204,8 @@ TEST_F(NVFuserTest, FusionAmpereViewMatmulTN_CUDA) {
 
   // Make warp tile:
   // -------------------------------------------------------------------------
-  scheduler_utils::matmul_utils::scheduleWarpTileWithReduction(tv2c, gemm_tile);
-  scheduler_utils::matmul_utils::scheduleWarpTileWithNoReduction(
-      tv2, gemm_tile);
+  mma_util::scheduleWarpTileWithReduction(tv2c, gemm_tile);
+  mma_util::scheduleWarpTileWithNoReduction(tv2, gemm_tile);
   //           -8  -7  -6  -5 -4 -3 -2 -1
   // [Mo No Ko Kwo Mwo Nwo Mw Nw Mi Ni Ki]
   tv0cr->computeAt(tv2c, -4);
@@ -2243,10 +2217,8 @@ TEST_F(NVFuserTest, FusionAmpereViewMatmulTN_CUDA) {
   tv0cw->merge(-2);
   tv0r->merge(-2);
   tv0_reshape->merge(-2);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0cw, gemm_tile, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0r, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0r, gemm_tile, 8);
   tv0cw->setMemoryType(MemoryType::Shared);
   // [Mo,Ko,i,wy,wx,v]
 
@@ -2254,10 +2226,8 @@ TEST_F(NVFuserTest, FusionAmpereViewMatmulTN_CUDA) {
   tv1cw->merge(-2);
   tv1r->merge(-2);
   // [No,Ko,i,wy,wx,v]
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1cw, gemm_tile, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1r, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1r, gemm_tile, 8);
   tv1cw->setMemoryType(MemoryType::Shared);
   // Schedule mma input
   // ---------------------------------------------------------------------------
@@ -2386,15 +2356,14 @@ TEST_F(NVFuserTest, FusionVoltaMatmulTNCrossWarp_CUDA) {
 
   // Make warp tile:
   // -------------------------------------------------------------------------
-  scheduler_utils::matmul_utils::scheduleWarpTileWithReduction(tv2c, gemm_tile);
+  mma_util::scheduleWarpTileWithReduction(tv2c, gemm_tile);
   auto tv2c_rf = tv2c->rFactor({-9, -4, -1});
 
   // tv2c_rf is the actual output of the mma op after
   //  Rfactoring.
   mma_builder.accumulatorTv(tv2c_rf);
 
-  scheduler_utils::matmul_utils::scheduleWarpTileWithNoReduction(
-      tv2, gemm_tile);
+  mma_util::scheduleWarpTileWithNoReduction(tv2, gemm_tile);
 
   //           -8   -7  -6 -5 -4 -3 -2 -1
   // [Mo No Ko Mwo  Nwo Kwo Mw Nw Mi Ni Ki]
@@ -2411,10 +2380,8 @@ TEST_F(NVFuserTest, FusionVoltaMatmulTNCrossWarp_CUDA) {
   // [Mo,No,Ko,N,M,K]
   tv0cw->merge(-2);
   tv0r->merge(-2);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0cw, gemm_tile, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0r, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0r, gemm_tile, 8);
   tv0cw->setMemoryType(MemoryType::Shared);
   // [Mo,Ko,i,wy,wx,v]
 
@@ -2422,10 +2389,8 @@ TEST_F(NVFuserTest, FusionVoltaMatmulTNCrossWarp_CUDA) {
   tv1cw->merge(-2);
   tv1r->merge(-2);
   // [Mo,No,Ko,i,wy,wx,v]
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1cw, gemm_tile, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1r, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1r, gemm_tile, 8);
   tv1cw->setMemoryType(MemoryType::Shared);
   // Schedule mma input
   // ---------------------------------------------------------------------------
@@ -2551,7 +2516,7 @@ TEST_F(NVFuserTest, FusionVoltaMatmulTNCrossCTA_CUDA) {
 
   // Make warp tile:
   // -------------------------------------------------------------------------
-  scheduler_utils::matmul_utils::scheduleWarpTileWithReduction(tv2c, gemm_tile);
+  mma_util::scheduleWarpTileWithReduction(tv2c, gemm_tile);
   //              -9 -8  -7  -6  -5 -4 -3 -2 -1
   // [Mo No K2CTA Ko Kwo Mwo Nwo Mw Nw Mi Ni Ki]
   auto tv2c_rf = tv2c->rFactor({-9, -8, -1});
@@ -2560,8 +2525,7 @@ TEST_F(NVFuserTest, FusionVoltaMatmulTNCrossCTA_CUDA) {
   //  Rfactoring.
   mma_builder.accumulatorTv(tv2c_rf);
 
-  scheduler_utils::matmul_utils::scheduleWarpTileWithNoReduction(
-      tv2, gemm_tile);
+  mma_util::scheduleWarpTileWithNoReduction(tv2, gemm_tile);
 
   //                 -8  -7  -6  -5 -4 -3 -2 -1
   // [Mo No K2CTA Ko Kwo Mwo Nwo Mw Nw Mi Ni Ki]
@@ -2578,10 +2542,8 @@ TEST_F(NVFuserTest, FusionVoltaMatmulTNCrossCTA_CUDA) {
   // [Mo,No,Ko,N,M,K]
   tv0cw->merge(-2);
   tv0r->merge(-2);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0cw, gemm_tile, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv0r, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv0r, gemm_tile, 8);
   tv0cw->setMemoryType(MemoryType::Shared);
   // [Mo,Ko,i,wy,wx,v]
 
@@ -2589,10 +2551,8 @@ TEST_F(NVFuserTest, FusionVoltaMatmulTNCrossCTA_CUDA) {
   tv1cw->merge(-2);
   tv1r->merge(-2);
   // [Mo,No,Ko,i,wy,wx,v]
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1cw, gemm_tile, 8);
-  scheduler_utils::matmul_utils::scheduleContiguousVectorLoad(
-      tv1r, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1cw, gemm_tile, 8);
+  mma_util::scheduleContiguousVectorLoad(tv1r, gemm_tile, 8);
   tv1cw->setMemoryType(MemoryType::Shared);
   // Schedule mma input
   // ---------------------------------------------------------------------------
@@ -2722,9 +2682,8 @@ TEST_F(NVFuserTest, FusionAmpereMatmulTNSwizzled_CUDA) {
 
   // Make warp tile:
   //
-  scheduler_utils::matmul_utils::scheduleWarpTileWithReduction(tv2c, gemm_tile);
-  scheduler_utils::matmul_utils::scheduleWarpTileWithNoReduction(
-      tv2, gemm_tile);
+  mma_util::scheduleWarpTileWithReduction(tv2c, gemm_tile);
+  mma_util::scheduleWarpTileWithNoReduction(tv2, gemm_tile);
   //           -8   -7 -6 -5 -4 -3 -2 -1
   // [Mo No Ko Kwo Mwo Nwo Mw Nw Mi Ni Ki]
   tv0cr->computeAt(tv2c, -4);
